requirements.txt
pip>=23.0
streamlit==1.33.0
PyPDF2==3.0.1
sentence-transformers==2.7.0 # Pinning latest minor version
torch==2.3.0
numpy==1.26.4
transformers==4.39.3
langchain-text-splitters==0.0.1
accelerate==0.28.0
protobuf>=3.20.3 # Optional, but recommended for compatibility
rank_bm25-0.2.2